feature(policiesinassembly): Add capabilities to workflow step
- Workflow step extracts files into temp folder
- Uses rules in all tests
- Also introduce a policies.name property to make distinguishing
folders easier.
