--- antenna-workflow-steps/processors/checkers/antenna-drools-checker/src/main/resources/bindings.xjb
@@ -1,24 +1,0 @@
-<!--
-  ~ Copyright (c) Bosch Software Innovations GmbH 2019.
-  ~
-  ~ All rights reserved. This program and the accompanying materials
-  ~ are made available under the terms of the Eclipse Public License v2.0
-  ~ which accompanies this distribution, and is available at
-  ~ http://www.eclipse.org/legal/epl-v20.html
-  ~
-  ~ SPDX-License-Identifier: EPL-2.0
-  -->
-<jxb:bindings version="1.0"
-              xmlns:jxb="http://java.sun.com/xml/ns/jaxb"
-              xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-              xmlns:ci="http://jaxb.dev.java.net/plugin/code-injector">
-
-    <jxb:bindings schemaLocation="policy.xsd">
-        <jxb:bindings node="//xsd:complexType[@name='policy']">
-            <jxb:class name="FromXmlPolicy" ref="org.eclipse.sw360.antenna.bundle.DroolsEvaluationResult"/>
-        </jxb:bindings>
-        <jxb:bindings node="//xsd:complexType[@name='severity']">
-            <jxb:class name="FromXmlSeverity" ref="org.eclipse.sw360.antenna.api.IEvaluationResult.Severity"/>
-        </jxb:bindings>
-    </jxb:bindings>
-</jxb:bindings>
