--- antenna-workflow-steps/processors/checkers/antenna-drools-checker/src/test/resources/policies/rules/DummyRule.drl
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) Bosch Software Innovations GmbH 2019.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-
-package org.eclipse.sw360.antenna.exampleproject
-
-import org.eclipse.sw360.antenna.model.artifact.Artifact;
-import org.eclipse.sw360.antenna.model.xml.generated.MatchState;
-import org.eclipse.sw360.antenna.bundle.DroolsEvaluationResult
-
-import java.util.List;
-
-rule "Artifacts must be open source"
-no-loop true
-when
-    a : Artifact( isProprietary() == true )
-    e : DroolsEvaluationResult( getId() == "Dummy" )
-then
-    modify (e) { addFailedArtifact(a) };
-end
