--- antenna-workflow-steps/processors/checkers/antenna-drools-checker/src/test/resources/policiesTheSecond/policies.properties
@@ -1,12 +1,0 @@
-#
-# Copyright (c) Bosch Software Innovations GmbH 2019.
-#
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v20.html
-#
-# SPDX-License-Identifier: EPL-2.0
-#
-
-policies.version=0.0.0
