+++ antenna-workflow-steps/processors/validators/antenna-drools-checker/src/main/java/org/eclipse/sw360/antenna/drools/DroolsRuleFolderVisitor.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) Bosch Software Innovations GmbH 2019.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+package org.eclipse.sw360.antenna.drools;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.ArrayList;
+import java.util.List;
+
+public class DroolsRuleFolderVisitor implements FileVisitor<Path> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(DroolsEngine.class);
+    private List<File> ruleFiles = new ArrayList<>();
+
+    public List<File> getRuleFiles() {
+        return ruleFiles;
+    }
+
+    @Override
+    public FileVisitResult preVisitDirectory(Path path, BasicFileAttributes basicFileAttributes) {
+        return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult visitFile(Path path, BasicFileAttributes basicFileAttributes) {
+        if (Files.isRegularFile(path) && path.toString().endsWith(".drl")) {
+            ruleFiles.add(path.toFile());
+        }
+        return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult visitFileFailed(Path path, IOException e) {
+        LOGGER.warn("File [" + path.toString() + "] could not be read.");
+        return FileVisitResult.CONTINUE;
+    }
+
+    @Override
+    public FileVisitResult postVisitDirectory(Path path, IOException e) {
+        return FileVisitResult.CONTINUE;
+    }
+}
