+++ antenna-workflow-steps/processors/validators/antenna-drools-checker/src/main/resources/policy.xsd
