+++ antenna-workflow-steps/processors/validators/antenna-drools-checker/src/test/resources/policies/policies.properties
@@ -9,5 +9,5 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 
-policies.name=Antenna Basic Rules
-policies.version=0.0.1
+policies.name=policies
+policies.version=0.0.0
