+++ antenna-workflow-steps/processors/validators/antenna-drools-checker/src/test/resources/policiesTheSecond/policies.xml
