+++ antenna-workflow-steps/processors/validators/antenna-drools-checker/src/test/resources/policiesTheSecond/rules/DummyRuleTheSecond.drl
@@ -9,20 +9,19 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.sw360.antenna.rules
+package org.eclipse.sw360.antenna.exampleproject
 
 import org.eclipse.sw360.antenna.model.artifact.Artifact;
 import org.eclipse.sw360.antenna.model.xml.generated.MatchState;
-import org.eclipse.sw360.antenna.drools.DroolsEvaluationResult;
+import org.eclipse.sw360.antenna.drools.DroolsEvaluationResult
 
 import java.util.List;
 
-rule "Unkown artifact"
+rule "Artifacts have to be open source"
 no-loop true
 when
-    a : Artifact( getMatchState() == MatchState.UNKNOWN,
-        isProprietary() == false )
-    e : DroolsEvaluationResult( getId() == "A1" )
+    a : Artifact( isProprietary() == true )
+    e : DroolsEvaluationResult( getId() == "DummyTheSecond" )
 then
     modify (e) { addFailedArtifact(a) };
 end
