+++ example-projects/example-policies/policies.properties
@@ -9,4 +9,5 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 
+policies.name=Test policies
 policies.version=0.0.0
