+++ example-projects/rules-test-development-project/src/main/resources/policies/policies.properties
@@ -9,4 +9,5 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 
+policies.name=Test rule development
 policies.version=0.0.0
